Reorganize tasks
