+++ core/src/main/kotlin/org/sert2521/sertain/subsystems/Subsystem.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.suspendCancellableCoroutine
+import org.sert2521.sertain.coroutines.RobotScope
 import org.sert2521.sertain.events.Use
 import org.sert2521.sertain.events.fire
 import kotlin.coroutines.coroutineContext
@@ -15,25 +16,10 @@ abstract class Subsystem(val name: String, val default: (suspend () -> Unit)? = 
 
     val occupied: Boolean
         get() = currentJob != null
-}
 
-suspend fun <R> use(
-    vararg subsystems: Subsystem,
-    cancelConflicts: Boolean = true,
-    name: String = "ANONYMOUS_TASK",
-    action: suspend CoroutineScope.() -> R
-): R {
-    val context = coroutineContext
-    return suspendCancellableCoroutine { continuation ->
-        CoroutineScope(context).launch {
-            fire(Use(
-                    subsystems.toSet(),
-                    cancelConflicts,
-                    name,
-                    context,
-                    continuation,
-                    action
-            ))
+    init {
+        RobotScope.launch {
+
         }
     }
 }
